# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Anketa API
  description: Social media voting platform — "Where the World Votes"
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local development

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # ─── Common ───────────────────────────────────────────────────────────────
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasNextPage:
          type: boolean
      required: [total, page, limit, hasNextPage]

    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
      required: [statusCode, message]

    # ─── Auth ─────────────────────────────────────────────────────────────────
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_]+$'
        password:
          type: string
          minLength: 8
      required: [email, username, password]

    LoginDto:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]

    RefreshDto:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]

    AuthTokensDto:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/UserDto'
      required: [accessToken, refreshToken, user]

    # ─── Users ────────────────────────────────────────────────────────────────
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        followersCount:
          type: integer
        followingCount:
          type: integer
        pollsCount:
          type: integer
        isFollowing:
          type: boolean
          description: Whether the authenticated user follows this user
        createdAt:
          type: string
          format: date-time
      required: [id, email, username, followersCount, followingCount, pollsCount, isFollowing, createdAt]

    UpdateProfileDto:
      type: object
      properties:
        bio:
          type: string
          maxLength: 200
        avatar:
          type: string
          format: uri

    # ─── Polls ────────────────────────────────────────────────────────────────
    OptionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        imageUrl:
          type: string
          nullable: true
        votesCount:
          type: integer
        percentage:
          type: number
          format: float
      required: [id, text, votesCount, percentage]

    PollDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        question:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        totalVotes:
          type: integer
        commentsCount:
          type: integer
        hasVoted:
          type: boolean
          description: Whether the authenticated user has voted on this poll
        userVotedOptionId:
          type: string
          format: uuid
          nullable: true
        author:
          $ref: '#/components/schemas/UserDto'
        options:
          type: array
          items:
            $ref: '#/components/schemas/OptionDto'
        createdAt:
          type: string
          format: date-time
      required: [id, question, totalVotes, commentsCount, hasVoted, author, options, createdAt]

    PollListDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PollDto'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required: [data, meta]

    CreatePollDto:
      type: object
      properties:
        question:
          type: string
          minLength: 5
          maxLength: 280
        description:
          type: string
          maxLength: 500
        category:
          type: string
        expiresAt:
          type: string
          format: date-time
        options:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
                minLength: 1
                maxLength: 100
              imageUrl:
                type: string
                format: uri
            required: [text]
          minItems: 2
          maxItems: 10
      required: [question, options]

    # ─── Votes ────────────────────────────────────────────────────────────────
    CastVoteDto:
      type: object
      properties:
        optionId:
          type: string
          format: uuid
      required: [optionId]

    VoteResultsDto:
      type: object
      properties:
        pollId:
          type: string
          format: uuid
        totalVotes:
          type: integer
        options:
          type: array
          items:
            $ref: '#/components/schemas/OptionDto'
      required: [pollId, totalVotes, options]

    # ─── Comments ─────────────────────────────────────────────────────────────
    CommentDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        author:
          $ref: '#/components/schemas/UserDto'
        createdAt:
          type: string
          format: date-time
      required: [id, content, author, createdAt]

    CommentListDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CommentDto'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      required: [data, meta]

    CreateCommentDto:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 500
      required: [content]

paths:
  # ─── Auth ───────────────────────────────────────────────────────────────────
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensDto'
        '409':
          description: Email or username already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensDto'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshDto'
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokensDto'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout (invalidate refresh token)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshDto'
      responses:
        '204':
          description: Logged out successfully

  # ─── Users ──────────────────────────────────────────────────────────────────
  /users/me:
    get:
      tags: [Users]
      summary: Get authenticated user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'

    patch:
      tags: [Users]
      summary: Update authenticated user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileDto'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'

  /users/{username}:
    get:
      tags: [Users]
      summary: Get user profile by username
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/{id}/follow:
    post:
      tags: [Users]
      summary: Follow a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Now following the user
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Already following
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Users]
      summary: Unfollow a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Unfollowed successfully
        '404':
          description: User not found or not following
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/{username}/polls:
    get:
      tags: [Users]
      summary: Get polls created by a user
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        '200':
          description: User's polls
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollListDto'

  # ─── Polls ──────────────────────────────────────────────────────────────────
  /polls:
    get:
      tags: [Polls]
      summary: Get trending polls (sorted by vote count)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Trending polls
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollListDto'

    post:
      tags: [Polls]
      summary: Create a new poll
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePollDto'
      responses:
        '201':
          description: Poll created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollDto'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /polls/following:
    get:
      tags: [Polls]
      summary: Get polls from followed users
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        '200':
          description: Following feed polls
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollListDto'

  /polls/{id}:
    get:
      tags: [Polls]
      summary: Get a poll by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Poll detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollDto'
        '404':
          description: Poll not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Polls]
      summary: Delete a poll (author only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Poll deleted
        '403':
          description: Not the poll author
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Poll not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Votes ──────────────────────────────────────────────────────────────────
  /polls/{id}/vote:
    post:
      tags: [Votes]
      summary: Cast a vote on a poll
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CastVoteDto'
      responses:
        '200':
          description: Vote cast, returns updated results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResultsDto'
        '409':
          description: Already voted on this poll
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: Poll has expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /polls/{id}/results:
    get:
      tags: [Votes]
      summary: Get vote results for a poll
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Poll vote results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResultsDto'

  # ─── Comments ───────────────────────────────────────────────────────────────
  /polls/{id}/comments:
    get:
      tags: [Comments]
      summary: Get comments on a poll
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: Poll comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListDto'

    post:
      tags: [Comments]
      summary: Add a comment to a poll
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentDto'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'

  /comments/{id}:
    delete:
      tags: [Comments]
      summary: Delete a comment (author only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Comment deleted
        '403':
          description: Not the comment author
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'